pyproject.toml
[tool.poetry]
name = "aiobot"
version = "0.1.0"
description = "Telegram BOT"
authors = ["Almaz Yessenyazov <mr.esenyazov@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
aiogram = "^2.13"
requests = "^2.25.1"
python-dotenv = "^0.17.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.4"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
